+++ src/app/search/search-filters.ts
@@ -16,6 +16,7 @@ import { D1StoresService } from '../inventory/d1-stores.service';
 import { D2StoresService } from '../inventory/d2-stores.service';
 import { querySelector } from '../shell/reducer';
 import { storesSelector } from '../inventory/reducer';
+import { maxLightLoadout } from '../loadout/auto-loadouts';
 
 /**
  * A selector for the search config for a particular destiny version.
@@ -153,6 +154,7 @@ export function buildSearchConfig(
     stackable: ['stackable'],
     category: Object.keys(categoryFilters),
     inloadout: ['inloadout'],
+    maxpower: ['maxpower'],
     new: ['new'],
     tag: ['tagged'],
     level: ['level'],
@@ -341,6 +343,7 @@ export function searchFilters(
   storeService: StoreServiceType
 ): SearchFilters {
   let _duplicates: { [hash: number]: DimItem[] } | null = null; // Holds a map from item hash to count of occurrances of that hash
+  const _maxPowerItems: string[] = [];
   let _lowerDupes = {};
   let _sortedStores: DimStore[] | null = null;
   let _loadoutItemIds: Set<string> | undefined;
@@ -453,6 +456,7 @@ export function searchFilters(
      */
     reset() {
       _duplicates = null;
+      _maxPowerItems.length = 0;
       _lowerDupes = {};
       _sortedStores = null;
     },
@@ -643,6 +647,19 @@ export function searchFilters(
       },
       masterwork(item: DimItem) {
         return item.masterwork;
+      },
+      maxpower(item: DimItem) {
+        if (!_maxPowerItems.length) {
+          storeService.getStores().forEach((store) => {
+            _maxPowerItems.push(
+              ..._.flatten(Object.values(maxLightLoadout(storeService, store).items)).map((i) => {
+                return i.id;
+              })
+            );
+          });
+        }
+
+        return _maxPowerItems.includes(item.id);
       },
       dupe(item: DimItem, predicate: string) {
         if (_duplicates === null) {