+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Removed a bunch of help popups.
 * Added information about unique stacks.
+* Added `is:maxpower` search to return highest light items.
 * Bounties with expiration times are now shown, and are sorted in front in order of expiration time.
 * Added masterwork tier range filter.
 * Highlight the stat that is boosted by masterwork in item details.